refactor: 코드 리팩토링 (FE-89)
